Fix: converting to webp for Info banner
