+++ next.config.mjs
@@ -3,6 +3,18 @@ const nextConfig = {
   compiler: {
     styledComponents: true,
   },
+  images: {
+    formats: ['image/avif', 'image/webp'],
+
+    remotePatterns: [
+      {
+        protocol: 'http',
+        hostname: 'localhost',
+        port: '3000',
+        pathname: '*',
+      },
+    ],
+  },
 }
 
 export default nextConfig
