+++ src/app/page.module.css
@@ -1,4 +1,5 @@
 .main {
+  margin: 0 auto;
   display: flex;
   flex-direction: column;
   justify-content: space-between;
