+++ src/widgets/Header/ui/styles.ts
@@ -4,11 +4,11 @@ import Link from 'next/link'
 
 const TheHeader = styled.header`
   // Приклеивает меню к верхнему краю страницы
-  // position: sticky;
-  // position: fixed;
+  position: sticky;
+  position: fixed;
+
   top: 0;
   right: 0;
-
   width: 100%;
   background-color: ${colors.backgroundBase2};
   color: ${colors.darkPrimary};
